requirements.txt
altair==5.5.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
awscli==1.42.45
blinker==1.9.0
boto3==1.40.43
botocore==1.40.45
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
docutils==0.19
dotenv==0.9.9
duckdb==1.4.0
executing==2.2.1
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
narwhals==2.7.0
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==21.0.0
pyasn1==0.6.1
PyAthena==3.18.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pyspark==4.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.7.2
s3transfer==0.14.0
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
xlrd==2.0.2
